opravena komunikacia i2c medzi motorom a stm32 a opravena komunikacia medzi raspberry a stm32, ziskavanie pozicie zatial neporiesene

+++ Nucleo/motor_master/src/Servo/Servo.h
@@ -10,6 +10,6 @@ void InitializePWMServo(void);
 
 void setServoRaw(uint32_t value);
 
-void setServo(double angle);
+void setServo(int16_t angle);
 
 #endif /* SERVO_H_ */
